\name{param.factor.bayes}
\alias{param.factor.bayes}
\title{Simulate Parameters for the Bayesian Logistic Regression...}
\description{Simulate Parameters for the Bayesian Logistic Regression}
\details{This method simulates parameters for the Bayesian Logistic Regression
(logit.bayes).}
\value{a list specifying link, link-inverse, random samples, and ancillary
parameters}
\arguments{\item{obj}{a zelig object}
\item{num}{an integer specifying the number of simulations to compute}
\item{...}{ignored parameters}}
