\name{paramMCMCZelig}
\alias{paramMCMCZelig}
\title{Simulate Parameters of All MCMC-Zelig Object...}
\usage{paramMCMCZelig(obj, num=1000, ...)}
\description{Simulate Parameters of All MCMC-Zelig Object}
\details{This method}
\value{a list representing a \code{parameters} object}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a "MCMCZelig" \code{zelig} object}
\item{num}{an integer specifying the number of simulations to produce for the
parameters}
\item{...}{ignored parameters}}
