\name{print.summary.MCMCZelig}
\alias{print.summary.MCMCZelig}
\title{Print a Summary MCMCZelig Object...}
\usage{print.summary.MCMCZelig(x, digits=max(3, getOption("digits") - 3), ...)}
\description{Print a Summary MCMCZelig Object}
\details{This method prints a summary object for \code{MCMCZelig} objects}
\value{a \code{summary.MCMCZelig} object}
\arguments{\item{object}{an "MCMCZelig" object}
\item{digits}{a numeric specifying the precision of the summary object}
\item{...}{ignored parameters}}
