\name{setx.factor.bayes}
\alias{setx.factor.bayes}
\title{Dummy setx Method for Bayesian Factor Regressions...}
\usage{setx.factor.bayes(obj, ...)}
\description{Dummy \code{setx} Method for Bayesian Factor Regressions}
\details{This method simply outputs an error message.}
\value{Nothing}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a "factor.bayes" MCMCZelig object}
\item{...}{ignored parameters}}
