\name{sim.factor.bayes}
\alias{sim.factor.bayes}
\title{Dummy sim Method for Bayesian Factor Regressions...}
\usage{sim.factor.bayes(obj, ...)}
\description{Dummy \code{sim} Method for Bayesian Factor Regressions}
\details{This method simply outputs an error message.}
\value{Nothing}
\author{Matt Owen \email{mowen@iq.harvard.edu}}
\arguments{\item{obj}{a "factor.bayes" MCMCZelig object}
\item{...}{ignored parameters}}
