\name{summary.MCMCZelig}
\alias{summary.MCMCZelig}
\title{Summary of MCMCZelig Object...}
\usage{summary.MCMCZelig(object, quantiles=c(0.025, 0.5, 0.975), ...)}
\description{Summary of MCMCZelig Object}
\details{This method produces a summary object for \code{MCMCZelig} objects}
\value{a \code{summary.MCMCZelig} object}
\arguments{\item{object}{an "MCMCZelig" object}
\item{quantiles}{a numeric vector specifying the quantiles to use in the
summary object.}}
