\name{zelig2factor.bayes}
\alias{zelig2factor.bayes}
\title{Interface between the Zelig Model factor...}
\usage{zelig2factor.bayes(formula, ..., data)}
\description{Interface between the Zelig Model factor.bayes and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
