\name{zelig2mlogit.bayes}
\alias{zelig2mlogit.bayes}
\title{Interface between the Zelig Model mlogit...}
\usage{zelig2mlogit.bayes(formula, ..., data)}
\description{Interface between the Zelig Model mlogit.bayes and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
