\name{zelig2oprobit.bayes}
\alias{zelig2oprobit.bayes}
\title{Interface between the Zelig Model oprobit...}
\usage{zelig2oprobit.bayes(formula, ..., data)}
\description{Interface between the Zelig Model oprobit.bayes and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
