\name{zelig2poisson.bayes}
\alias{zelig2poisson.bayes}
\title{Interface between the Zelig Model poisson...}
\usage{zelig2poisson.bayes(formula, ..., data)}
\description{Interface between the Zelig Model poisson.bayes and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
