\name{zelig2probit.bayes}
\alias{zelig2probit.bayes}
\title{Interface between the Zelig Model probit...}
\usage{zelig2probit.bayes(formula, ..., data)}
\description{Interface between the Zelig Model probit.bayes and 
the Pre-existing Model-fitting Method}
\value{a list specifying '.function'}
\arguments{\item{formula}{a formula}
\item{...}{additonal parameters}
\item{data}{a data.frame}}
