\name{ZeligBayesian-package}
\title{ZeligBayesian...}
\description{ZeligBayesian}
\details{\tabular{ll}{
Package: \tab ZeligBayesian\cr
Version: \tab 0.1\cr
Date: \tab 2011-04-25\cr
Depends: Zelig
License: \tab GPL version 2 or newer\cr
}

Edit this description}
\alias{ZeligBayesian-package}
\alias{ZeligBayesian}
\docType{package}
\author{Matt Owen}
\keyword{package}
