\name{print.summary.MCMCZelig}
\alias{print.summary.MCMCZelig}
\title{Print Summarized MCMC-Zelig Objects...}
\usage{print.summary.MCMCZelig(x, digits=max(3, getOption("digits") - 3), ...)}
\description{Print Summarized MCMC-Zelig Objects}
\details{This function prints the summarized MCMC-Zelig objects}
\value{Nothing}
\arguments{\item{x}{a \code{summary.MCMCZelig} object}
\item{digits}{an integer specifying the precision of decimal numbers}
\item{...}{ignored parameters}}
