\name{summary.MCMCZelig}
\alias{summary.MCMCZelig}
\title{Summarize MCMC-Zelig Objects...}
\usage{\method{summary}{MCMCZelig} (object, quantiles=c(0.025, 0.5, 0.975),
    ...)}
\description{Summarize MCMC-Zelig Objects}
\details{This function summarizes the MCMC-Zelig fitted statistical model}
\value{a \code{summary.MCMCZelig} Object}
\arguments{\item{object}{a \code{MCMCZelig} Object}
\item{quantiles}{the quantiles that are desired for output}
\item{...}{ignored parameters}}
